Fix structural lint failures and CodeRabbit issues for profile editor
